backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
email-validator>=2.1.0
pyjwt>=2.9.0
pydantic-settings>=2.6.0
openai-agents>=0.1.0
qdrant-client>=1.12.0
httpx>=0.28.0
python-dotenv>=1.0.0
bcrypt>=4.2.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
